2-OVSeg/SAN/requirements.txt
cython
scipy
shapely
timm
h5py
submitit
scikit-image
wandb
setuptools
numpy==1.22.4
Pillow==9.3.0
pycocotools~=2.0.4
fvcore
tabulate
tqdm
ftfy
regex
opencv-python
open_clip_torch==2.16.0
mmcv==1.6.0
gradio
huggingface_hub
shapely
future
omegaconf>=2.1
iopath==0.1.9
nuscenes-devkit